Fixed the solving of jan_22_2023 puzzle
