+++ tests/conftest.py
@@ -1,0 +1,21 @@
+import pytest
+from puzzle_service import app
+
+
+@pytest.fixture()
+def service():
+    # app.config.update({
+    #     "TESTING": True,
+    # })
+
+    # other setup can go here
+
+    yield app
+
+    # clean up / reset resources here
+
+
+@pytest.fixture()
+def client(service):
+    return app.test_client()
+
