+++ web/style.css
@@ -28,6 +28,12 @@ body {
   margin: 0 0 12px;
 }
 
+.puzzle_type {
+}
+
+.status_area {
+}
+
 .puzzle_control {
   margin-left: 10px;
   margin-right: 10px;
